Remove single partition protocols.
