+++ core/src/test/java/io/atomix/core/tree/impl/RaftDocumentTreeTest.java
@@ -16,7 +16,7 @@
 package io.atomix.core.tree.impl;
 
 import io.atomix.primitive.protocol.PrimitiveProtocol;
-import io.atomix.protocols.raft.RaftProtocol;
+import io.atomix.protocols.raft.MultiRaftProtocol;
 
 /**
  * Document tree test using the Raft protocol.
@@ -24,7 +24,7 @@ import io.atomix.protocols.raft.RaftProtocol;
 public class RaftDocumentTreeTest extends DocumentTreeTest {
   @Override
   protected PrimitiveProtocol protocol() {
-    return RaftProtocol.builder()
+    return MultiRaftProtocol.builder()
         .withMaxRetries(5)
         .build();
   }
