+++ protocols/primary-backup/src/main/java/io/atomix/protocols/backup/partition/PrimaryBackupPartitionGroupConfig.java
@@ -19,7 +19,7 @@ import io.atomix.primitive.partition.MemberGroupProvider;
 import io.atomix.primitive.partition.MemberGroupStrategy;
 import io.atomix.primitive.partition.PartitionGroupConfig;
 import io.atomix.primitive.protocol.PrimitiveProtocol;
-import io.atomix.protocols.backup.PrimaryBackupProtocol;
+import io.atomix.protocols.backup.MultiPrimaryProtocol;
 
 /**
  * Primary-backup partition group configuration.
@@ -31,7 +31,7 @@ public class PrimaryBackupPartitionGroupConfig extends PartitionGroupConfig<Prim
 
   @Override
   public PrimitiveProtocol.Type getType() {
-    return PrimaryBackupProtocol.TYPE;
+    return MultiPrimaryProtocol.TYPE;
   }
 
   @Override
