+++ protocols/primary-backup/src/main/resources/META-INF/services/io.atomix.primitive.protocol.PrimitiveProtocolFactory
@@ -13,5 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-io.atomix.protocols.backup.PrimaryBackupProtocolFactory
 io.atomix.protocols.backup.MultiPrimaryProtocolFactory
