+++ protocols/raft/src/main/java/io/atomix/protocols/raft/MultiRaftProtocolConfig.java
@@ -16,20 +16,25 @@
 package io.atomix.protocols.raft;
 
 import io.atomix.primitive.Recovery;
-import io.atomix.primitive.partition.Partitioner;
 import io.atomix.primitive.protocol.PrimitiveProtocol;
+import io.atomix.primitive.protocol.PrimitiveProtocolConfig;
 import io.atomix.protocols.raft.proxy.CommunicationStrategy;
 
 import java.time.Duration;
 import java.util.concurrent.Executor;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
 /**
- * Multi-Raft protocol configuration.
+ * Raft protocol configuration.
  */
-public class MultiRaftProtocolConfig extends RaftProtocolConfig {
-  private Partitioner<String> partitioner = Partitioner.MURMUR3;
+public class MultiRaftProtocolConfig extends PrimitiveProtocolConfig<MultiRaftProtocolConfig> {
+  private Duration minTimeout = Duration.ofMillis(250);
+  private Duration maxTimeout = Duration.ofSeconds(30);
+  private ReadConsistency readConsistency = ReadConsistency.SEQUENTIAL;
+  private CommunicationStrategy communicationStrategy = CommunicationStrategy.LEADER;
+  private Recovery recoveryStrategy = Recovery.RECOVER;
+  private int maxRetries = 0;
+  private Duration retryDelay = Duration.ofMillis(100);
+  private Executor executor;
 
   @Override
   public PrimitiveProtocol.Type getType() {
@@ -37,82 +42,172 @@ public class MultiRaftProtocolConfig extends RaftProtocolConfig {
   }
 
   /**
-   * Returns the protocol partitioner.
+   * Returns the minimum session timeout.
    *
-   * @return the protocol partitioner
+   * @return the minimum session timeout
    */
-  public Partitioner<String> getPartitioner() {
-    return partitioner;
+  public Duration getMinTimeout() {
+    return minTimeout;
   }
 
   /**
-   * Sets the protocol partitioner.
+   * Sets the minimum session timeout.
    *
-   * @param partitioner the protocol partitioner
+   * @param minTimeout the minimum session timeout
+   * @return the Raft protocol configuration
+   */
+  public MultiRaftProtocolConfig setMinTimeout(Duration minTimeout) {
+    this.minTimeout = minTimeout;
+    return this;
+  }
+
+  /**
+   * Returns the maximum session timeout.
+   *
+   * @return the maximum session timeout
+   */
+  public Duration getMaxTimeout() {
+    return maxTimeout;
+  }
+
+  /**
+   * Sets the maximum session timeout.
+   *
+   * @param maxTimeout the maximum session timeout
+   * @return the Raft protocol configuration
+   */
+  public MultiRaftProtocolConfig setMaxTimeout(Duration maxTimeout) {
+    this.maxTimeout = maxTimeout;
+    return this;
+  }
+
+  /**
+   * Returns the read consistency level.
+   *
+   * @return the read consistency level
+   */
+  public ReadConsistency getReadConsistency() {
+    return readConsistency;
+  }
+
+  /**
+   * Sets the read consistency level.
+   *
+   * @param readConsistency the read consistency level
+   * @return the Raft protocol configuration
+   */
+  public MultiRaftProtocolConfig setReadConsistency(ReadConsistency readConsistency) {
+    this.readConsistency = readConsistency;
+    return this;
+  }
+
+  /**
+   * Returns the client communication strategy.
+   *
+   * @return the client communication strategy
+   */
+  public CommunicationStrategy getCommunicationStrategy() {
+    return communicationStrategy;
+  }
+
+  /**
+   * Sets the client communication strategy.
+   *
+   * @param communicationStrategy the client communication strategy
+   * @return the Raft protocol configuration
+   */
+  public MultiRaftProtocolConfig setCommunicationStrategy(CommunicationStrategy communicationStrategy) {
+    this.communicationStrategy = communicationStrategy;
+    return this;
+  }
+
+  /**
+   * Returns the client recovery strategy.
+   *
+   * @return the client recovery strategy
+   */
+  public Recovery getRecoveryStrategy() {
+    return recoveryStrategy;
+  }
+
+  /**
+   * Sets the client recovery strategy.
+   *
+   * @param recoveryStrategy the client recovery strategy
+   * @return the Raft protocol configuration
+   */
+  public MultiRaftProtocolConfig setRecoveryStrategy(Recovery recoveryStrategy) {
+    this.recoveryStrategy = recoveryStrategy;
+    return this;
+  }
+
+  /**
+   * Returns the maximum allowed number of retries.
+   *
+   * @return the maximum allowed number of retries
+   */
+  public int getMaxRetries() {
+    return maxRetries;
+  }
+
+  /**
+   * Sets the maximum allowed number of retries.
+   *
+   * @param maxRetries the maximum allowed number of retries
    * @return the protocol configuration
    */
-  public MultiRaftProtocolConfig setPartitioner(Partitioner<String> partitioner) {
-    this.partitioner = checkNotNull(partitioner, "partitioner cannot be null");
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setGroup(String group) {
-    super.setGroup(group);
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setMinTimeout(Duration minTimeout) {
-    super.setMinTimeout(minTimeout);
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setMaxTimeout(Duration maxTimeout) {
-    super.setMaxTimeout(maxTimeout);
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setReadConsistency(ReadConsistency readConsistency) {
-    super.setReadConsistency(readConsistency);
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setCommunicationStrategy(CommunicationStrategy communicationStrategy) {
-    super.setCommunicationStrategy(communicationStrategy);
-    return this;
-  }
-
-  @Override
-  public MultiRaftProtocolConfig setRecoveryStrategy(Recovery recoveryStrategy) {
-    super.setRecoveryStrategy(recoveryStrategy);
-    return this;
-  }
-
-  @Override
   public MultiRaftProtocolConfig setMaxRetries(int maxRetries) {
-    super.setMaxRetries(maxRetries);
+    this.maxRetries = maxRetries;
     return this;
   }
 
-  @Override
+  /**
+   * Returns the retry delay.
+   *
+   * @return the retry delay
+   */
+  public Duration getRetryDelay() {
+    return retryDelay;
+  }
+
+  /**
+   * Sets the retry delay.
+   *
+   * @param retryDelayMillis the retry delay in milliseconds
+   * @return the protocol configuration
+   */
   public MultiRaftProtocolConfig setRetryDelayMillis(long retryDelayMillis) {
-    super.setRetryDelayMillis(retryDelayMillis);
-    return this;
+    return setRetryDelay(Duration.ofMillis(retryDelayMillis));
   }
 
-  @Override
+  /**
+   * Sets the retry delay.
+   *
+   * @param retryDelay the retry delay
+   * @return the protocol configuration
+   */
   public MultiRaftProtocolConfig setRetryDelay(Duration retryDelay) {
-    super.setRetryDelay(retryDelay);
+    this.retryDelay = retryDelay;
     return this;
   }
 
-  @Override
+  /**
+   * Returns the executor.
+   *
+   * @return the executor
+   */
+  public Executor getExecutor() {
+    return executor;
+  }
+
+  /**
+   * Sets the executor.
+   *
+   * @param executor the executor
+   * @return the protocol configuration
+   */
   public MultiRaftProtocolConfig setExecutor(Executor executor) {
-    super.setExecutor(executor);
+    this.executor = executor;
     return this;
   }
 }
