+++ protocols/raft/src/main/java/io/atomix/protocols/raft/partition/RaftPartitionGroupConfig.java
@@ -17,7 +17,7 @@ package io.atomix.protocols.raft.partition;
 
 import io.atomix.primitive.partition.PartitionGroupConfig;
 import io.atomix.primitive.protocol.PrimitiveProtocol;
-import io.atomix.protocols.raft.RaftProtocol;
+import io.atomix.protocols.raft.MultiRaftProtocol;
 import io.atomix.storage.StorageLevel;
 
 import java.util.HashSet;
@@ -37,7 +37,7 @@ public class RaftPartitionGroupConfig extends PartitionGroupConfig<RaftPartition
 
   @Override
   public PrimitiveProtocol.Type getType() {
-    return RaftProtocol.TYPE;
+    return MultiRaftProtocol.TYPE;
   }
 
   @Override
